builder/requirements.txt
marko~=2.0.0
pyyaml~=6.0.1
jinja2~=3.1.2
marshmallow~=3.20.1
